requirements.txt
Flask==3.0.0
flask-cors==4.0.0
flask-limiter==3.5.0
flask-swagger-ui==4.11.1
flask-compress==1.14
sqlalchemy==2.0.23
alembic==1.13.0
pyjwt==2.8.0
marshmallow==3.20.1
Werkzeug==3.0.1
pandas==2.2.0
matplotlib==3.8.2
seaborn==0.13.1
numpy==1.26.3
